+++ app/code/local/Autocompleteplus/Autosuggest/Model/Mysql4/Pusher.php
@@ -1,13 +1,9 @@
 <?php
 
-class Autocompleteplus_Autosuggest_Model_Mysql4_Pusher extends Mage_Core_Model_Resource_Db_Abstract
+/**
+ * For Magento < 1.6 backward compatibility
+ */
+class Autocompleteplus_Autosuggest_Model_Mysql4_Pusher extends Autocompleteplus_Autosuggest_Model_Resource_Pusher
 {
-    protected function _construct() {
-        $this->_init('autocompleteplus_autosuggest/pusher', 'id');
-    }
-
-    public function truncate() {
-        $this->_getWriteAdapter()->query('TRUNCATE TABLE '.$this->getMainTable());
-        return $this;
-    }
+    
 }