+++ app/code/local/Autocompleteplus/Autosuggest/Model/Resource/Batches/Collection.php
@@ -1,6 +1,6 @@
 <?php
 
-class Autocompleteplus_Autosuggest_Model_Resource_Batches_Collection extends Mage_Core_Model_Resource_Db_Collection_Abstract
+class Autocompleteplus_Autosuggest_Model_Resource_Batches_Collection extends Mage_Core_Model_Mysql4_Collection_Abstract
 {
     /**
      * Initialize resource collection.