+++ app/code/local/Autocompleteplus/Autosuggest/Model/Resource/Checksum.php
@@ -1,6 +1,6 @@
 <?php
 
-class Autocompleteplus_Autosuggest_Model_Resource_Checksum extends Mage_Core_Model_Resource_Db_Abstract
+class Autocompleteplus_Autosuggest_Model_Resource_Checksum extends Mage_Core_Model_Mysql4_Abstract
 {
     /**
      * Resource initialization.