+++ app/code/local/Autocompleteplus/Autosuggest/Model/Resource/Notifications.php
@@ -13,7 +13,7 @@
  * @copyright  Copyright (c) 2014 Fast Simon (http://www.instantsearchplus.com)
  * @license    http://opensource.org/licenses/osl-3.0.php  Open Software License (OSL 3.0)
  */
-class Autocompleteplus_Autosuggest_Model_Resource_Notifications extends Mage_Core_Model_Resource_Db_Abstract
+class Autocompleteplus_Autosuggest_Model_Resource_Notifications extends Mage_Core_Model_Mysql4_Abstract
 {
     protected function _construct()
     {