+++ app/code/local/Autocompleteplus/Autosuggest/Model/Resource/Notifications/Collection.php
@@ -14,7 +14,7 @@
  * @license    http://opensource.org/licenses/osl-3.0.php  Open Software License (OSL 3.0)
  */
 class Autocompleteplus_Autosuggest_Model_Resource_Notifications_Collection
-    extends Mage_Core_Model_Resource_Db_Collection_Abstract
+    extends Mage_Core_Model_Mysql4_Collection_Abstract
 {
     protected function _construct()
     {