+++ app/code/local/Autocompleteplus/Autosuggest/controllers/LayeredController.php
@@ -19,16 +19,10 @@ class Autocompleteplus_Autosuggest_LayeredController extends Mage_Core_Controlle
     {
         $response = $this->getResponse();
         $request = $this->getRequest();
-        
-        $helper = Mage::helper('autocompleteplus_autosuggest');
         $authkey = $request->getParam('authentication_key');
         $uuid = $request->getParam('uuid');
-        
         $scope = $request->getParam('scope', 'stores');
-        if ($scope != 'stores' && $scope != 'default' && $scope != ''){
-            $scope = 'stores';
-        }
-        $scopeId = $helper->validateInput($request->getParam('store_id', Mage::app()->getStore()->getStoreId()), 'integer', 1, null);
+        $scopeId = $request->getParam('store_id', 1);
 
         if (!$this->valid($uuid, $authkey)) {
             $resp = json_encode(array('status' => 'error: '.'Authentication failed'));
@@ -60,16 +54,11 @@ class Autocompleteplus_Autosuggest_LayeredController extends Mage_Core_Controlle
     {
         $request = $this->getRequest();
         $response = $this->getResponse();
-        $helper = Mage::helper('autocompleteplus_autosuggest');
         $authkey = $request->getParam('authentication_key');
         $uuid = $request->getParam('uuid');
-        
         $scope = $request->getParam('scope', 'stores');
-        if ($scope != 'stores' && $scope != 'default' && $scope != ''){
-            $scope = 'stores';
-        }
-        $scopeId = $helper->validateInput($request->getParam('store_id', Mage::app()->getStore()->getStoreId()), 'integer', 1, null);
-        
+        $scopeId = $request->getParam('store_id', 1);
+
         if (!$this->valid($uuid, $authkey)) {
             $resp = json_encode(array('status' => 'error: '.'Authentication failed'));
 
@@ -101,11 +90,10 @@ class Autocompleteplus_Autosuggest_LayeredController extends Mage_Core_Controlle
     {
         $request = $this->getRequest();
         $response = $this->getResponse();
-        $helper = Mage::helper('autocompleteplus_autosuggest');
 
         $authkey = $request->getParam('authentication_key');
         $uuid = $request->getParam('uuid');
-        $scopeId = $helper->validateInput($request->getParam('store_id', Mage::app()->getStore()->getStoreId()), 'integer', 1, null);
+        $scopeId = $request->getParam('store_id', 1);
 
         if (!$this->valid($uuid, $authkey)) {
             $resp = json_encode(array('status' => $this->__('error: Authentication failed')));